requirements.txt
python-telegram-bot==13.7
schedule==1.1.0
requests==2.26.0
huggingface-hub==0.0.19